AutoReconnect support for MongoEngine
Summary:
AutoReconnect happens when pymongo loses contact with the current
primary in a replica set but will try to re-establish the connection for
the next pymongo request. This happens when the primary changes.

On all reads, if we get an AutoReconnect exception, we'll wait 100ms and
retry without erroring out. Should stop (or limit) the brief error
spikes we get when we make a replica set member stand down.

Test Plan:
Setup a replica set locally and swapped primary and secondary back and
forth as I refreshed various pages. Grepped for _collection and _cursor
references in QuerySet.

Reviewers: josh

Manifest Tasks: T2139