Change auto-indexing semantics
New semantics are that index creation is background by default and will
only create if QuerySet._allow_index_creation is True. If it's False,
it'll raise an exception.

We only set allow_index_creation to True for dev & stage. This should
cause a hard error starting the server in testing if we missed creating
an index in production and won't block stage whenever someone adds an
index to an existing collection.