MongoEngine query parser to replace QuerySet
Summary:
Improved what I did at Hackathon to make a working query parser. It's
not completely perfect yet (there are some cases of obscure queries it
probably won't handle right, but I don't think we have any of these in
our code).

For the first little while, as we convert queries to use this, we should
validate that they work (and, if they don't, we should write a test case
for it and fix the bug). But, the parser's pretty conservative, so most
failures should raise exceptions instead of silently fail.

Also adds the ability to easily do a query on secondaries.

Test Plan: Unit tests

Reviewers: josh

Differential Revision: http://infra1.corp.contextlogic.com/D6412