Automatically convert strings to ObjectIds in $in
Summary:
Description:
To maintain backwards compatibility (and deal with odd bugs), if a
string is passed to something in the context of an ObjectIdField,
convert it to an ObjectId.

This commit fixes handling for that case with $in where you want to
convert a list of strings to a list of ObjectIds.

Test Plan: Unit tests & by_ids() works now

Reviewers: josh

Differential Revision: http://infra1.corp.contextlogic.com/D6425