Change number of retries / time for AutoReconnect
Summary:
In prod, sometimes we need a bit more than 100ms for mongos to come back
or replica set election to finish and we want to handle those errors
gracefully. Instead of 1 retry of 100ms, we do up to 4 at 75ms each
(300ms total). So, worst-case we're not significantly delaying requests,
but we get a lower error rate.

Test Plan:
Made the changes on a single FE in prod, did some mongos bounces and
couldn't trigger an AutoReconnect exception anymore.

Reviewers: jack, josh, david

Maniphest Tasks: T2699

Differential Revision: http://infra1.corp.contextlogic.com/D9742