Disable AutoReconnect retries for now
Summary:
I think this is the cause of T6724. Since we retry up to 5 times at 75ms
wait each time and async mongo means we can have 50-100 connections
open, the FE can get chewed up in sleep() for tens of seconds which
explains what we're seeing in prod.

Disabling this for now and will try tonight to see what happens. If that
fixes the problem, then I'll build an async way of doing the sleep.

Test Plan:
Forced AutoReconnect to happen by restarting a machine in a stage
environment. Ran a benchmark during the restart activity and 99th%
latency was cut in half.

Reviewers: josh

Maniphest Tasks: T6724

Differential Revision: http://infra1.corp.contextlogic.com/D11276