Ensure that 'id' is present and first in '_fields_ordered' for all classes
An 'id' field only needs to be explicitly added to '_fields_ordered' by
the TopLevelDocumentMetaclass if 'id_field' is not provided and the 'id'
field was created in the metaclass. Otherwise, the field would have been
added to '_fields_ordered' along with the other user-defined fields. Thus,
the conditional for explicitly adding 'id' to '_fields_ordered' can be
moved inwards in scope.

More importantly, 'id' should always be the first element of
'_fields_ordered'. This may not always occur if a Document class is
derived from a non-abstract Document class. This commit ensures that 'id'
will remain the first element for all classes.