Use NEAREST read pref instead of SECONDARY_PREFERRED
Summary:
Nearest read pref allows queries to be routed to either primary or
secondary (vs secondary preferred, which'll pick a secondary with worse
tag sets over a primary).

Need to do this so we can use the serving-primary tagset to force reads
to go to 01.

Test Plan: Did some queries

Reviewers: david, kevin, andrew, greg

Maniphest Tasks: T20959

Differential Revision: https://phab.wish.com/D28869