Raise TypeError when __in-query used with Document
Doing a query like `BlogPost.objects(authors__in=author).count()`, where
`author` is a single Document, will now raise a TypeError explaining
that you cannot use a single `Document` with the `in`-operator.
i.e. `BlogPost.objects(authors__in=[author]).count()`

Similarly using the `in`-operator with a non-iterable will also raise a
TypeError explaining that you need to use an iterable with the `in`-operator.

The same applies to the `nin`, `all`, and `near` operators.

Note:
  The added code uses `_import_class('BaseDocument')` to break any circular
  dependency.