Adding greenlet sync primitives and ioloop periodic executor
Summary:
Needed for pymongo 3 support.
We'll have to monkey patch out a few things in pymongo, one of which
being the periodic executor and topology classes. Pymongo3 has support for
custom condition classes and monitor classes, but the topology class makes its
own threading.Lock, which won't work with greenlets.

The periodic executor implements all of the same methods as the one in pymongo,
but throws out //all// of the code because it's designed for threads, so
there's no point subclasseing.

This is only part of what's needed, but I wanted to get eyes on the stuff
that's mongo-independent before patching it in.

Test Plan: unit tests, included

Reviewers: adam, david

Maniphest Tasks: T18386

Differential Revision: https://phab.wish.com/D33410