Make bool(d) = True for all Document/EmbeddedDocument
Summary:
EmbeddedDocuments without any fields had a __len__ of 0, making
__nonzero__ = False (i.e. bool(d) = False) for those documents.

When we're trying to do a null check, `if d` should succeed if `d` is a
Document or EmbeddedDocument, even if it's an empty one since what we're
trying to determine is whether or not `d` is `None`.

Test Plan: Wrote unit test, it failed. Added `__nonzero__`, it passed.

Reviewers: david, jim, andrew

Subscribers:

Maniphest Tasks: T25086

Differential Revision: https://phab.wish.com/D33944