Log if upsert and multi are both True
Summary: Mgo doesn't support both update multi AND upsert in the same update. We could just upsert after a multi update if we didn't modify any documents, but grepping the codebase nothing seems to use the behavior. We'll verify that with this logger, and raise an exception on the mongoproxy side and punt it for later if we end up needing it anywhere

Test Plan: did an update call, ran unit tests

Reviewers: bpalsson, tjackson, david, rli

Differential Revision: https://phab.wish.com/D46021