Disallow passing in fields/excluded fields when skip_transform is specified in find_and_modify
Summary:
fields/excluded fields is how we unmarshal the bson documents into python objects.

It would take a bunch of duplication to get it to work correctly, but no one uses it right now so it's not worth the effort

the proper solution is just to replace skip_transform with a proper integration for find_and_modify that supports all the operators that we want to use.

Test Plan: local testing

Reviewers: bpalsson, tjackson, shaw, david, rli

Differential Revision: https://phab.wish.com/D49300