Improve save() performance by skipping index creation
Indexes are now only created when a Model is first used, eg the first
call to save() or on the first call to _get_collection(), or when the new
meta["auto_create_index_on_save"] option is set to True. This is a minor
breaking change for some applications. As a workaround apps can explicitly
call ensure_indexes() or set meta["auto_create_index_on_save"] to True.